resources/fastapi/pyproject.toml
[tool.poetry]
name = "ppt-generation"
version = "0.1.0"
description = ""
authors = ["sauravniraula <sauravniraula912@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.4"
fastapi = { extras = ["standard"], version = "^0.115.6" }
langchain-core = "^0.3.28"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.13"
langchain-text-splitters = "^0.3.4"
pymupdf = "^1.25.1"
pymupdf4llm = "^0.0.17"
pillow = "^11.0.0"
supabase = "^2.10.0"
langchain-google-genai = "^2.0.7"
python-pptx = "^1.0.2"
boto3 = "1.35.90"
replicate = "^1.0.4"
pdf2image = "^1.17.0"
pydub = "^0.25.1"
resend = "^2.5.1"
watchtower = "^3.3.1"
python-docx = "^1.1.2"
google-genai = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"